Fixed some bugs I was to tired to test I guess lol
