+++ think_big_2023/src/helpers.py
@@ -1,23 +1,6 @@
 import asyncio
 import sys
 from bot import bot
-import aiohttp
-
-_session = None
-
-
-def get_session():
-    global _session
-    if _session is None:
-        _session = aiohttp.ClientSession()
-    return _session
-
-
-def close_session():
-    global _session
-    if _session is not None:
-        _session.close()
-        _session = None
 
 
 def export(func):
